+++ TestSupport/EfSchemeCompare/Internal/CompareHelpers.cs
@@ -1,6 +1,7 @@
 ﻿// Copyright (c) 2017 Jon P Smith, GitHub: JonPSmith, web: http://www.thereformedprogrammer.net/
 // Licensed under MIT licence. See License.txt in the project root for license information.
 
+using System;
 using System.Collections.Generic;
 using System.Linq;
 using Microsoft.EntityFrameworkCore;
@@ -17,9 +18,12 @@ namespace TestSupport.EfSchemeCompare.Internal
             return FormSchemaTable(eRel.Schema, eRel.TableName);
         }
 
-        public static string FormSchemaTable(this DatabaseTable table)
+        public static string FormSchemaTable(this DatabaseTable table, string defaultSchema)
         {
-            return FormSchemaTable(table.Schema, table.Name);
+            //The DatabaseTable always provides a schema name, while the database Model provides null if default schema name.
+            //This makes sure that name will match the EF Core Model format
+            var schemaToUse = table.Schema == defaultSchema ? null : table.Schema;
+            return FormSchemaTable(schemaToUse, table.Name);
         }
 
         public static string NullableAsString(this bool isNullable)
@@ -27,16 +31,35 @@ namespace TestSupport.EfSchemeCompare.Internal
             return isNullable ? "NULL" : "NOT NULL";
         }
 
+        public static StringComparison GetStringComparison(this StringComparer caseComparer)
+        {
+            return ComparerToComparison[caseComparer];
+        }
+
+        private static readonly Dictionary<StringComparer, StringComparison> ComparerToComparison =
+            new Dictionary<StringComparer, StringComparison>
+            {
+                {StringComparer.CurrentCulture, StringComparison.CurrentCulture},
+                {StringComparer.CurrentCultureIgnoreCase, StringComparison.CurrentCultureIgnoreCase},
+                {StringComparer.InvariantCulture, StringComparison.InvariantCulture},
+                {StringComparer.InvariantCultureIgnoreCase, StringComparison.InvariantCultureIgnoreCase},
+                {StringComparer.Ordinal, StringComparison.Ordinal},
+                {StringComparer.OrdinalIgnoreCase, StringComparison.OrdinalIgnoreCase}
+            };
+
         public static string CombinedColNames(this IEnumerable<IProperty> properties)
         {
             return string.Join(",", properties.Select(x => x.Relational().ColumnName));
         }
 
-        //The scaffold does not set the correct ValueGenerated for a column that has a sql default value
-        public static string ConvertNullableValueGenerated(this ValueGenerated? valGen, string sqlDefault)
+        //The scaffold does not set the correct ValueGenerated for a column that has a sql default value of a computed column
+        //see https://github.com/aspnet/EntityFrameworkCore/issues/9323
+        public static string ConvertNullableValueGenerated(this ValueGenerated? valGen, string computedColumnSql, string defaultValueSql)
         {
-            if (valGen == null && sqlDefault != null)
+            if (valGen == null && defaultValueSql != null)
                 return ValueGenerated.OnAdd.ToString();
+            if (valGen == null && computedColumnSql != null)
+                return ValueGenerated.OnAddOrUpdate.ToString();
             return valGen?.ToString() ?? ValueGenerated.Never.ToString();
         }