+++ TestSupport/EfSchemeCompare/Internal/Stage2Comparer.cs
@@ -1,6 +1,7 @@
 ﻿// Copyright (c) 2017 Jon P Smith, GitHub: JonPSmith, web: http://www.thereformedprogrammer.net/
 // Licensed under MIT licence. See License.txt in the project root for license information.
 
+using System;
 using System.Collections.Generic;
 using System.Collections.Immutable;
 using System.Linq;
@@ -11,6 +12,7 @@ namespace TestSupport.EfSchemeCompare.Internal
     internal class Stage2Comparer
     {
         private readonly DatabaseModel _databaseModel;
+        private readonly StringComparer _caseComparer;
 
         private bool _hasErrors;
 
@@ -18,10 +20,11 @@ namespace TestSupport.EfSchemeCompare.Internal
         private readonly IReadOnlyList<CompareLog> _ignoreList;
         public IReadOnlyList<CompareLog> Logs => _logs.ToImmutableList();
 
-        public Stage2Comparer(DatabaseModel databaseModel, IReadOnlyList<CompareLog> ignoreList = null)
+        public Stage2Comparer(DatabaseModel databaseModel, CompareEfSqlConfig config = null)
         {
             _databaseModel = databaseModel;
-            _ignoreList = ignoreList ?? new List<CompareLog>();
+            _ignoreList = config?.LogsToIgnore ?? new List<CompareLog>();
+            _caseComparer = config?.CaseComparer ?? StringComparer.CurrentCulture;
         }
 
         public bool CompareLogsToDatabase(IReadOnlyList<CompareLog> firstStageLogs)
@@ -38,11 +41,11 @@ namespace TestSupport.EfSchemeCompare.Internal
         private void LookForUnusedTables(IReadOnlyList<CompareLog> firstStageLogs, CompareLog log)
         {
             var logger = new CompareLogger(CompareType.Table, null, log.SubLogs, _ignoreList, () => _hasErrors = true);
-            var databaseTableNames = _databaseModel.Tables.Select(x => x.Name);
+            var databaseTableNames = _databaseModel.Tables.Select(x => x.FormSchemaTable(_databaseModel.DefaultSchema));
             var allEntityTableNames = firstStageLogs.SelectMany(p => p.SubLogs)
                 .Where(x => x.State == CompareState.Ok && x.Type == CompareType.Entity)
                 .Select(p => p.Expected).OrderBy(p => p).Distinct().ToList();
-            var tablesNotUsed = databaseTableNames.Where(p => !allEntityTableNames.Contains(p));
+            var tablesNotUsed = databaseTableNames.Where(p => !allEntityTableNames.Contains(p, _caseComparer));
 
             foreach (var tableName in tablesNotUsed)
             {
@@ -52,15 +55,15 @@ namespace TestSupport.EfSchemeCompare.Internal
 
         private void LookForUnusedColumns(IReadOnlyList<CompareLog> firstStageLogs, CompareLog log)
         {
-            var logger = new CompareLogger(CompareType.Column, null, log.SubLogs, _ignoreList, () => _hasErrors = true);
-            var tableDict = _databaseModel.Tables.ToDictionary(x => x.Name);
+            var logger = new CompareLogger(CompareType.Column, null, _logs, _ignoreList, () => _hasErrors = true);
+            var tableDict = _databaseModel.Tables.ToDictionary(x => x.FormSchemaTable(_databaseModel.DefaultSchema), _caseComparer);
             //because of table splitting and TPH we need to groups properties by table name to correctly say what columns are missed
             var entityColsGrouped = firstStageLogs.SelectMany(p => p.SubLogs)
                 .Where(x => x.State == CompareState.Ok && x.Type == CompareType.Entity)
                 .GroupBy(x => x.Expected, y => y.SubLogs
                     .Where(x => x.State == CompareState.Ok && x.Type == CompareType.Property)
                     .Select(p => p.Expected));
-            var entityColsByTableDict = entityColsGrouped.ToDictionary(x => x.Key, y => y.SelectMany(x => x.ToList()));
+            var entityColsByTableDict = entityColsGrouped.ToDictionary(x => x.Key, y => y.SelectMany(x => x.ToList()), _caseComparer);
 
             foreach (var entityLog in firstStageLogs.SelectMany(p => p.SubLogs)
                 .Where(x => x.State == CompareState.Ok && x.Type == CompareType.Entity))
@@ -68,7 +71,7 @@ namespace TestSupport.EfSchemeCompare.Internal
                 if (tableDict.ContainsKey(entityLog.Expected))
                 {
                     var dbColNames = tableDict[entityLog.Expected].Columns.Select(x => x.Name);
-                    var colsNotUsed = dbColNames.Where(p => !entityColsByTableDict[entityLog.Expected].Contains(p));
+                    var colsNotUsed = dbColNames.Where(p => !entityColsByTableDict[entityLog.Expected].Contains(p, _caseComparer));
                     foreach (var colName in colsNotUsed)
                     {
                         logger.ExtraInDatabase(colName, CompareAttributes.ColumnName, entityLog.Expected);
@@ -79,8 +82,8 @@ namespace TestSupport.EfSchemeCompare.Internal
 
         private void LookForUnusedIndexes(IReadOnlyList<CompareLog> firstStageLogs, CompareLog log)
         {
-            var logger = new CompareLogger(CompareType.Index, null, log.SubLogs, _ignoreList, () => _hasErrors = true);
-            var tableDict = _databaseModel.Tables.ToDictionary(x => x.Name);
+            var logger = new CompareLogger(CompareType.Index, null, _logs, _ignoreList, () => _hasErrors = true);
+            var tableDict = _databaseModel.Tables.ToDictionary(x => x.FormSchemaTable(_databaseModel.DefaultSchema), _caseComparer);
             foreach (var entityLog in firstStageLogs.SelectMany(p => p.SubLogs)
                 .Where(x => x.State == CompareState.Ok && x.Type == CompareType.Entity))
             {
@@ -90,7 +93,7 @@ namespace TestSupport.EfSchemeCompare.Internal
                     var allEfIndexNames = entityLog.SubLogs
                         .Where(x => x.State == CompareState.Ok && x.Type == CompareType.Index)
                         .Select(p => p.Expected).OrderBy(p => p).Distinct().ToList();
-                    var indexesNotUsed = indexCol.Where(p => !allEfIndexNames.Contains(p));
+                    var indexesNotUsed = indexCol.Where(p => !allEfIndexNames.Contains(p, _caseComparer));
                     foreach (var indexName in indexesNotUsed)
                     {
                         logger.ExtraInDatabase(indexName, CompareAttributes.IndexConstraintName, entityLog.Expected);